changed canvas_and_display example to canvas_and_gr
